write filteredimage endpoint

+++ course-02/project/image-filter-starter-code/src/util/util.ts
@@ -8,17 +8,26 @@ import Jimp = require('jimp');
 //    inputURL: string - a publicly accessible url to an image file
 // RETURNS
 //    an absolute path to a filtered image locally saved file
-export async function filterImageFromURL(inputURL: string): Promise<string>{
-    return new Promise( async resolve => {
-        const photo = await Jimp.read(inputURL);
-        const outpath = '/tmp/filtered.'+Math.floor(Math.random() * 2000)+'.jpg';
-        await photo
-        .resize(256, 256) // resize
-        .quality(60) // set JPEG quality
-        .greyscale() // set greyscale
-        .write(__dirname+outpath, (img)=>{
-            resolve(__dirname+outpath);
+export async function filterImageFromURL(inputURL: string): Promise<string> {
+    let photo: Jimp;
+    try {
+        photo = await Jimp.read(inputURL);
+    } catch (error) {
+        console.error("Could not read the file from the link provided.");
+        return new Promise((resolve, reject) => {
+            return reject("Could not read the file from the link provided.")
         });
+    }
+
+    return new Promise(async resolve => {
+        const outpath = '/tmp/filtered.' + Math.floor(Math.random() * 2000) + '.jpg';
+        await photo
+            .resize(256, 256) // resize
+            .quality(60) // set JPEG quality
+            .greyscale() // set greyscale
+            .write(__dirname + outpath, (img) => {
+                resolve(__dirname + outpath);
+            });
     });
 }
 
@@ -27,8 +36,19 @@ export async function filterImageFromURL(inputURL: string): Promise<string>{
 // useful to cleanup after tasks
 // INPUTS
 //    files: Array<string> an array of absolute paths to files
-export async function deleteLocalFiles(files:Array<string>){
-    for( let file of files) {
-        fs.unlinkSync(file);
+export async function deleteLocalFiles(files: Array<string>) {
+    for (let file of files) {
+        if (file) {
+            try {
+                fs.unlinkSync(file);
+            } catch (error) {
+                console.error("Error deleting the file : ", error);
+            }
+        }
+
     }
+}
+
+export function getTempFiles(currFile: string) {
+    return fs.readdirSync(__dirname + '/tmp').filter(file => currFile !== __dirname + '/tmp/' + file).map(file => __dirname + '/tmp/' + file);
 }
